2024-05-pycon-us-logfire/pyproject.toml
[project]
name = "logfire-talk"
version = "0.1.0"
description = "Code snippets from talk on Pydantic Logfire."
authors = [
    { name = "Samuel Colvin", email = "s@muelcolvin.com" }
]
dependencies = [
    "fastapi>=0.110.0",
    "logfire[fastapi,httpx,asyncpg,system-metrics,requests,openai]>=0.46.1",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.28.0",
    "asyncpg>=0.29.0",
    "python-multipart>=0.0.9",
    "openai>=1.14.1",
    "pillow>=10.2.0",
    "stripe>=9.6.0",
    "pandas>=2.2.2",
    "beautifulsoup4>=4.12.3",
    "fastui>=0.6.0",
    "pydantic-settings>=2.3.4",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.ruff]
line-length = 120
target-version = "py312"
lint.extend-select = ["Q", "RUF100", "C90", "UP", "I"]
lint.flake8-quotes = {inline-quotes = "single", multiline-quotes = "double"}
lint.mccabe = { max-complexity = 14 }
format.quote-style = "single"
